refactor(core): separate AmendTask from CreateTask commands
