+++ src/domain/task/command/amendtask.ts
@@ -1,0 +1,8 @@
+import type { TaskStatus } from '../entity/task'
+
+export type AmendTask<I = string> = {
+  readonly id: I
+  readonly timestamp: Date
+  readonly status?: TaskStatus
+  readonly messageKey?: string
+}
