+++ src/domain/task/command/createTask.ts
@@ -8,10 +8,3 @@ export type CreateTask<I = string, T = string> = {
   readonly messageKey: string
   readonly initiator?: string
 }
-
-export type AmendTask<I = string> = {
-  readonly id: I
-  readonly timestamp: Date
-  readonly status?: TaskStatus
-  readonly messageKey?: string
-}
